some bugs fixed
